+++ CHANGELOG.md
@@ -6,6 +6,21 @@
 
 _Nothing just yet_
 
+## [1.4.0] - 2026-07-25
+
+### Added
+
+- CalOPPA/CCPA/GDPR privacy policy
+- Policy change email workflow
+
+### Changed
+
+- Database backups are now encrypted
+
+### Fixed
+
+- Raw HTML tags appear in suggestion card titles
+
 ## [1.3.0] - 2026-07-25
 
 ### Added
@@ -158,7 +173,8 @@ This is a development version bump. Most of the changes are related to refactori
 - Toggle themes based on Richard Linklater's filmography
 - Toggle light and dark mode
 
-[Unreleased]: https://github.com/nschneble/linklater/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/nschneble/linklater/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/nschneble/linklater/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/nschneble/linklater/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/nschneble/linklater/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/nschneble/linklater/compare/v1.1.0...v1.2.0